@@ -6,7 +6,7 @@ import { Component, OnInit } from '@angular/core';
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
-export class AppComponent implements OnInit{
+export class AppComponent implements OnInit {
   title = 'Dating app ';
   usuarios: any;
 
@@ -19,7 +19,4 @@ export class AppComponent implements OnInit{
       complete: () => console.log('Request realizado com sucesso.')
     })
   }
-
-
-
 }
